Move traits to separate crate
