+++ Cargo.toml
@@ -18,6 +18,9 @@ rand = { version = "0.3.8", optional = true }
 rustc-serialize = { version = "0.3.13", optional = true }
 serde = { version = "^0.7.0", optional = true }
 
+[dependencies.num-traits]
+path = "./traits"
+
 [dev-dependencies]
 # Some tests of non-rand functionality still use rand because the tests
 # themselves are randomized.
